feat: updated some stuff

+++ src/events/guild/GuildDelete.ts
@@ -10,8 +10,8 @@ class GuildDelete extends Listener {
   public override async execute(guild: Guild) {
     const { database } = this.client;
 
-    const data = await database.query("SELECT guild_id FROM guilds WHERE guild_id = ?", [guild.id]);
-    if (data.length !== 0) {
+    const data = await database.query("SELECT guild_id FROM guilds WHERE guild_id = ?", [guild.id]).catch(() => void 0);
+    if (data && data.length !== 0) {
       await database.query("DELETE FROM guilds WHERE guild_id = ?", [guild.id]);
     }
